renaming configuration directory to configuration_change
