--- solutions/configuration/CLUS22_usecase/interfaces.yaml
@@ -1,23 +1,0 @@
-interfaces:
-  - name: GigabitEthernet1
-    type: ethernetCsmacd
-    enabled: "true"
-    description: MANAGEMENT INTERFACE - DON'T TOUCH ME
-    ipv4:
-      ip: 10.10.20.48
-      subnet: 255.255.255.0
-  - name: GigabitEthernet2
-    type: ethernetCsmacd
-    enabled: "false"
-    description: "Network Interface"
-  - name: GigabitEthernet3
-    type: ethernetCsmacd
-    enabled: "false"
-    description: "Network Interface"
-  - name: Loopback1
-    type: softwareLoopback
-    enabled: "true"
-    description: "Configured from Python"
-    ipv4:
-      ip: 10.20.50.20
-      subnet: 255.255.255.0
