--- solutions/configuration/CLUS22_usecase/template.j2
@@ -1,23 +1,0 @@
-{
-    "ietf-interfaces:interfaces": {
-        "interface": [
-            {% for interface in interfaces %}
-            {
-                "name": "{{ interface.name }}",
-                "description": "{{ interface.description }}",
-                "type": "iana-if-type:{{ interface.type }}",
-                "enabled": {{ interface.enabled }}
-                {% if "ipv4" in interface %},
-                "ietf-ip:ipv4": {
-                    "address": [
-                        {
-                            "ip": "{{ interface.ipv4.ip }}",
-                            "netmask": "{{ interface.ipv4.subnet }}"
-                        }
-                    ]
-                }{% endif %}
-            }{{ "," if not loop.last else "" }}
-            {% endfor %}
-        ]
-    }
-}
